Refactor worker pool and concurrency limiter implementations.
Reorganized and renamed worker pool and concurrency controls to improve clarity and modularity. Replaced `Limiter` with `SlotGroup` and simplified worker pool API with `WorkerPool` and `WorkFunc`. Updated related tests and removed unnecessary abstractions.
