+++ concurrency/slotgroup.go
@@ -5,10 +5,10 @@ import (
 	"sync/atomic"
 )
 
-// Limiter is a concurrency control mechanism that enforces a maximum level of parallelism.
+// SlotGroup is a control mechanism that enforces a maximum level of parallelism.
 // It provides a fixed number of "slots" that can be acquired and released by workers.
 // Thread-safe mechanisms are used to ensure safe access and prevent resource contention.
-type Limiter struct {
+type SlotGroup struct {
 
 	// slots is a buffered channel that represents the available worker slots.
 	// Workers acquire slots from this channel to execute tasks and release them afterward.
@@ -25,11 +25,11 @@ type Limiter struct {
 //region Implementation
 
 // Acquire returns the slots channel, allowing workers to acquire available slots for execution.
-// Workers should read from this channel before starting work to ensure that the concurrency
+// Workers should read from this channel before starting work to ensure that the conclimiter
 // limit is not exceeded.
 // This method is non-blocking as it only returns the channel.
-func (l *Limiter) Acquire() <-chan struct{} {
-	return l.slots
+func (sg *SlotGroup) Acquire() <-chan struct{} {
+	return sg.slots
 }
 
 // Release returns a slot back to the limiter, making it available for other workers.
@@ -37,19 +37,19 @@ func (l *Limiter) Acquire() <-chan struct{} {
 // been closed (in which case the release is ignored entirely).
 //
 // Thread-safe and ensures no deadlocks occur when releasing slots.
-func (l *Limiter) Release() {
+func (sg *SlotGroup) Release() {
 
 	// Do nothing if already closed.
-	if l.isClosed() {
+	if sg.isClosed() {
 		return
 	}
 
-	// Prevent blocking if the l.slots channel's buffer is already full.
-	if len(l.slots) == cap(l.slots) {
+	// Prevent blocking if the sg.slots channel's buffer is already full.
+	if len(sg.slots) == cap(sg.slots) {
 		return
 	}
 
-	l.slots <- struct{}{}
+	sg.slots <- struct{}{}
 }
 
 // Close shuts down the limiter, preventing further acquisitions and releases.
@@ -58,22 +58,22 @@ func (l *Limiter) Release() {
 //
 // This method is idempotent and prevents race conditions using a mutex.
 // Once closed, the limiter cannot be reused.
-func (l *Limiter) Close() {
-	l.closeMutex.Lock()
-	defer l.closeMutex.Unlock()
+func (sg *SlotGroup) Close() {
+	sg.closeMutex.Lock()
+	defer sg.closeMutex.Unlock()
 
 	// Prevent close from being run twice
-	if l.isClosed() {
+	if sg.isClosed() {
 		return
 	}
 
 	// Drain the slots channel of any remaining objects
-	for range len(l.slots) {
-		<-l.slots
+	for range len(sg.slots) {
+		<-sg.slots
 	}
 
-	l.closed.Store(true)
-	close(l.slots)
+	sg.closed.Store(true)
+	close(sg.slots)
 }
 
 //endregion
@@ -82,27 +82,27 @@ func (l *Limiter) Close() {
 
 // isClosed is a helper method that checks whether the limiter has been closed.
 // Returns true if the atomic `closed` flag is set.
-func (l *Limiter) isClosed() bool {
-	return l.closed.Load() == true
+func (sg *SlotGroup) isClosed() bool {
+	return sg.closed.Load() == true
 }
 
 //endregion
 
 //region Constructor
 
-// NewLimiter initializes a new Limiter instance with a specified number of slots.
-// The size parameter defines the maximum concurrency level, controlling the number of
+// NewSlotGroup initializes a new SlotGroup instance with a specified number of slots.
+// The size parameter defines the maximum conclimiter level, controlling the number of
 // slots that workers can acquire concurrently.
 //
-// Returns a fully initialized Limiter with the specified size.
-func NewLimiter(size uint16) *Limiter {
+// Returns a fully initialized SlotGroup with the specified size.
+func NewSlotGroup(size uint16) *SlotGroup {
 
 	slotBuffer := make(chan struct{}, size)
 	for i := uint16(0); i < size; i++ {
 		slotBuffer <- struct{}{}
 	}
 
-	return &Limiter{
+	return &SlotGroup{
 		slots:      slotBuffer,
 		closed:     &atomic.Bool{},
 		closeMutex: &sync.Mutex{},
