+++ concurrency/slotgroup_test.go
@@ -6,19 +6,19 @@ import (
 	"time"
 )
 
-func TestLimiter_AcquireRelease(t *testing.T) {
+func TestSlotGroup_AcquireRelease(t *testing.T) {
 	size := uint16(3)
-	limiter := NewLimiter(size)
+	slotGroup := NewSlotGroup(size)
 
 	// Validate that the slots channel is initialized with the correct capacity.
-	if cap(limiter.Acquire()) != int(size) {
-		t.Fatalf("expected limiter to have capacity %d, but got %d", size, cap(limiter.Acquire()))
+	if cap(slotGroup.Acquire()) != int(size) {
+		t.Fatalf("expected SlotGroup to have capacity %d, but got %d", size, cap(slotGroup.Acquire()))
 	}
 
 	// Acquire all slots and ensure that the channel behaves as expected.
 	for i := 0; i < int(size); i++ {
 		select {
-		case <-limiter.Acquire():
+		case <-slotGroup.Acquire():
 			// Slot successfully acquired
 		default:
 			t.Fatalf("could not acquire slot %d", i+1)
@@ -27,9 +27,9 @@ func TestLimiter_AcquireRelease(t *testing.T) {
 
 	// Release slots and ensure they are reusable.
 	for i := 0; i < int(size); i++ {
-		limiter.Release()
+		slotGroup.Release()
 		select {
-		case <-limiter.Acquire():
+		case <-slotGroup.Acquire():
 			// Slot successfully reacquired
 		default:
 			t.Fatalf("could not reacquire slot %d", i+1)
@@ -37,34 +37,34 @@ func TestLimiter_AcquireRelease(t *testing.T) {
 	}
 }
 
-func TestLimiter_ReleaseFullChannel(t *testing.T) {
+func TestSlotGroup_ReleaseFullChannel(t *testing.T) {
 	size := uint16(2)
-	limiter := NewLimiter(size)
+	slotGroup := NewSlotGroup(size)
 
 	// Fully acquire all slots.
 	for i := 0; i < int(size); i++ {
-		<-limiter.Acquire()
+		<-slotGroup.Acquire()
 	}
 
 	// Attempt to release more slots than the capacity and ensure no panic occurs.
 	for i := 0; i < 5; i++ {
-		limiter.Release()
+		slotGroup.Release()
 	}
 }
 
-func TestLimiter_Close(t *testing.T) {
+func TestSlotGroup_Close(t *testing.T) {
 	size := uint16(2)
-	limiter := NewLimiter(size)
+	slotGroup := NewSlotGroup(size)
 
 	// Acquire a slot to simulate active usage.
-	<-limiter.Acquire()
+	<-slotGroup.Acquire()
 
-	// Close the limiter.
-	limiter.Close()
+	// Close the SlotGroup.
+	slotGroup.Close()
 
 	// Ensure the slots channel is properly closed.
 	select {
-	case _, ok := <-limiter.Acquire():
+	case _, ok := <-slotGroup.Acquire():
 		if ok {
 			t.Fatal("expected slots channel to be closed after Close()")
 		}
@@ -73,15 +73,15 @@ func TestLimiter_Close(t *testing.T) {
 	}
 
 	// Ensure no panic occurs when calling Release after closure.
-	limiter.Release()
+	slotGroup.Release()
 
 	// Ensure no duplicate closure occurs.
-	limiter.Close()
+	slotGroup.Close()
 }
 
-func TestLimiter_ConcurrentAccess(t *testing.T) {
+func TestSlotGroup_ConcurrentAccess(t *testing.T) {
 	size := uint16(5)
-	limiter := NewLimiter(size)
+	slotGroup := NewSlotGroup(size)
 
 	// Use a WaitGroup to coordinate concurrent goroutines.
 	var wg sync.WaitGroup
@@ -92,8 +92,8 @@ func TestLimiter_ConcurrentAccess(t *testing.T) {
 		go func() {
 			defer wg.Done()
 			select {
-			case <-limiter.Acquire():
-				limiter.Release()
+			case <-slotGroup.Acquire():
+				slotGroup.Release()
 			default:
 				// No slot available
 			}
@@ -103,10 +103,10 @@ func TestLimiter_ConcurrentAccess(t *testing.T) {
 	// Wait for all goroutines to complete.
 	wg.Wait()
 
-	// Ensure that all slots are reusable after concurrency.
+	// Ensure that all slots are reusable after SlotGroup use.
 	for i := 0; i < int(size); i++ {
 		select {
-		case <-limiter.Acquire():
+		case <-slotGroup.Acquire():
 			// Slot successfully acquired
 		default:
 			t.Fatalf("could not acquire slot %d after concurrent access", i+1)
@@ -114,38 +114,38 @@ func TestLimiter_ConcurrentAccess(t *testing.T) {
 	}
 }
 
-func TestLimiter_CloseWhileInUse(t *testing.T) {
+func TestSlotGroup_CloseWhileInUse(t *testing.T) {
 	size := uint16(2)
-	limiter := NewLimiter(size)
+	slotGroup := NewSlotGroup(size)
 
 	// Acquire one slot to simulate active usage.
-	<-limiter.Acquire()
+	<-slotGroup.Acquire()
 
-	// Close the limiter while slots are still being used.
+	// Close the SlotGroup while slots are still being used.
 	go func() {
 		time.Sleep(10 * time.Millisecond)
-		limiter.Close()
+		slotGroup.Close()
 	}()
 
 	// Simulate a short delay for testing cleanup behavior.
 	time.Sleep(50 * time.Millisecond)
 
 	// Ensure no panic occurs with closure during usage.
-	limiter.Release()
+	slotGroup.Release()
 }
 
-func TestLimiter_ReleaseAfterClose(t *testing.T) {
+func TestSlotGroup_ReleaseAfterClose(t *testing.T) {
 	size := uint16(3)
-	limiter := NewLimiter(size)
+	slotGroup := NewSlotGroup(size)
 
-	// Close the limiter.
-	limiter.Close()
+	// Close the SlotGroup.
+	slotGroup.Close()
 
 	// Release after closure to ensure no panic occurs and no slot is reused.
-	limiter.Release()
+	slotGroup.Release()
 
 	select {
-	case _, ok := <-limiter.Acquire():
+	case _, ok := <-slotGroup.Acquire():
 		if ok {
 			t.Fatal("expected slots channel to be closed, but it remains open")
 		}
