--- worker/func.go
@@ -1,3 +1,0 @@
-package worker
-
-type Func[INPUT any] func(INPUT) error
