--- worker/worker.go
@@ -1,5 +1,0 @@
-package worker
-
-type Worker[INPUT any] interface {
-	Execute(INPUT) error
-}
