+++ workerpool/workfunc.go
@@ -1,0 +1,3 @@
+package workerpool
+
+type WorkFunc[INPUT any] func(INPUT) error
